# doc/API仕様書/BFF/openapi.yaml
openapi: 3.1.0
info:
  title: Nextjs Sample - BFF API
  version: 1.0.0
  description: |
    Next.js が公開する BFF (Backend for Frontend) の API 契約。
    ブラウザ (フロントエンド) がこのアプリ自身の `/api/**` に対して
    呼び出すエンドポイントを定義する。実装は `src/app/api/**` の Route Handlers。
servers:
  - url: /api
    description: このNext.jsアプリ自身が提供するBFFのエンドポイント
tags:
  - name: health
    description: 稼働確認
  - name: user
    description: ユーザー
  - name: todo
    description: Todoの管理
paths:
  /health:
    get:
      tags: [health]
      summary: ヘルスチェック
      description: サービスの稼働確認用エンドポイント。認証不要で常に200を返す。
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok

  /users:
    get:
      tags: [user]
      summary: ユーザー一覧取得
      description: 登録されている全ユーザーを返す。ページングは行わない。
      operationId: listUsers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "../common/schemas/User.yaml#/User"

  /todos:
    get:
      tags: [todo]
      summary: Todo一覧取得
      description: 登録されている全Todoを追加順(古い順)で返す。
      operationId: listTodos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "../common/schemas/Todo.yaml#/Todo"
    post:
      tags: [todo]
      summary: Todo作成
      description: 新しいTodoを登録する。title・dueDate・assignee・statusが必須。
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../common/schemas/Todo.yaml#/TodoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "../common/schemas/Todo.yaml#/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"

  /todos/{id}:
    put:
      tags: [todo]
      summary: Todo更新
      description: 指定したidのTodoをtitle・dueDate・assignee・statusで全体更新する。
      operationId: updateTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../common/schemas/Todo.yaml#/TodoInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "../common/schemas/Todo.yaml#/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [todo]
      summary: Todo削除
      description: 指定したidのTodoを削除する。対象が存在しない場合も204を返す(冪等)。
      operationId: deleteTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content

components:
  responses:
    BadRequest:
      description: 400 Bad Request
      content:
        application/json:
          schema:
            $ref: "../common/schemas/Error.yaml#/Error"
    NotFound:
      description: 404 Not Found
      content:
        application/json:
          schema:
            $ref: "../common/schemas/Error.yaml#/Error"